MovieTools: extract download code to separate, mockable class
