+++ app/src/main/java/com/battlelancer/seriesguide/util/SgPicassoRequestHandler.kt
@@ -11,6 +11,7 @@ import android.content.Context
 import android.net.Uri
 import androidx.core.net.toUri
 import com.battlelancer.seriesguide.SgApp
+import com.battlelancer.seriesguide.movies.tools.MoviePosterDownloader
 import com.battlelancer.seriesguide.tmdbapi.TmdbTools
 import com.battlelancer.seriesguide.tmdbapi.TmdbTools2
 import com.squareup.picasso.Downloader
@@ -67,8 +68,8 @@ class SgPicassoRequestHandler(
 
             val posterPath: String? = try {
                 runBlocking {
-                    SgApp.getServicesComponent(context).movieTools()
-                        .getMoviePosterPath(movieTmdbId)
+                    val tmdbMovies = SgApp.getServicesComponent(context).moviesService()
+                    MoviePosterDownloader(context, tmdbMovies).getMoviePosterPath(movieTmdbId)
                 }
             } catch (e: InterruptedException) {
                 null // Do nothing
